+++ Rider/Editor/Properties/AssemblyInfo.cs
@@ -4,7 +4,7 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyTitle("Unity.Rider.Editor")]
 [assembly: InternalsVisibleTo("Unity.Rider.EditorTests")]
 [assembly: InternalsVisibleTo("Unity.PackageValidationSuite.Editor")]
-[assembly: InternalsVisibleTo("Assembly-CSharp-Editor")]
+[assembly: InternalsVisibleTo("CSProjTest")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
 
 [assembly: AssemblyVersion("3.0.7")]