+++ Rider/Editor/RiderInitializer.cs
@@ -1,7 +1,5 @@
-using System;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using Packages.Rider.Editor.Util;
 using Rider.Editor.Util;
 using UnityEngine;
@@ -16,7 +14,7 @@ namespace Packages.Rider.Editor
         var assembly = EditorPluginInterop.EditorPluginAssembly;
         if (EditorPluginInterop.EditorPluginIsLoadedFromAssets(assembly))
         {
-          Debug.LogError($"Please delete {assembly.Location}. Unity 2019.2+ loads it directly from Rider installation. To disable this, open Rider's settings, search and uncheck 'Automatically install and update Rider's Unity editor plugin'.");
+          Debug.LogError($"Please delete {assembly.GetLoadedAssemblyLocation()}. Unity 2019.2+ loads it directly from Rider installation. To disable this, open Rider's settings, search and uncheck 'Automatically install and update Rider's Unity editor plugin'.");
           return;
         }
         
@@ -51,7 +49,7 @@ namespace Packages.Rider.Editor
           if (dllFile.Exists)
           {
             var bytes = File.ReadAllBytes(dllFile.FullName); 
-            assembly = AppDomain.CurrentDomain.Load(bytes); // doesn't lock assembly on disk
+            assembly = AssemblyLoading.LoadFromBytes(bytes); // doesn't lock assembly on disk
             if (PluginSettings.SelectedLoggingLevel >= LoggingLevel.TRACE)
               Debug.Log($"Rider EditorPlugin loaded from {dllFile.FullName}");
           
@@ -91,7 +89,7 @@ namespace Packages.Rider.Editor
           return;
         }
 
-        var assembly = AppDomain.CurrentDomain.Load(AssemblyName.GetAssemblyName(dllFile.FullName));
+        var assembly = AssemblyLoading.LoadFromPath(dllFile.FullName);
         if (PluginSettings.SelectedLoggingLevel >= LoggingLevel.TRACE)
           Debug.Log($"Rider EditorPlugin loaded from {dllFile.FullName}");